Move all parquet init code to core
